requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.4.0
ccxt==4.5.17
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coincurve==20.0.0
cryptography==46.0.3
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
numpy==2.3.4
pandas==2.3.3
propcache==0.4.1
psycopg2-binary==2.9.11
pycares==4.11.0
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
